+++ inlang/source-code/cli/src/commands/project/index.ts
@@ -1,8 +1,0 @@
-import { Command } from "commander"
-import { validate } from "./validate.js"
-
-export const project = new Command()
-	.command("project")
-	.description("Commands for managing your inlang project")
-	.argument("[command]")
-	.addCommand(validate)