+++ inlang/source-code/ide-extension/src/decorations/messagePreview.ts
@@ -4,6 +4,7 @@ import { contextTooltip } from "./contextTooltip.js"
 import { getStringFromPattern } from "../utilities/query.js"
 import { getActiveTextEditor } from "../utilities/initProject.js"
 import { CONFIGURATION } from "../configuration.js"
+import { resolveEscapedCharacters } from "../utilities/resolveEscapedCharacters.js"
 
 const MAXIMUM_PREVIEW_LENGTH = 40
 
@@ -21,8 +22,8 @@ export async function messagePreview(args: { context: vscode.ExtensionContext })
 			return
 		}
 
-		// TODO: this is a hack to prevent the message preview from showing up in the project.inlang.json file
-		if (activeTextEditor.document.fileName.includes("project.inlang.json")) {
+		// TODO: this is a hack to prevent the message preview from showing up in the project.inlang/settings file
+		if (activeTextEditor.document.fileName.includes("project.inlang")) {
 			return activeTextEditor.setDecorations(messagePreview, [])
 		}
 
@@ -50,7 +51,7 @@ export async function messagePreview(args: { context: vscode.ExtensionContext })
 
 				const variant = _message?.variants?.find((v) => v.languageTag === sourceLanguageTag)
 
-				const translation = getStringFromPattern({
+				const translationString = getStringFromPattern({
 					pattern: variant?.pattern || [
 						{
 							type: "Text",
@@ -60,6 +61,8 @@ export async function messagePreview(args: { context: vscode.ExtensionContext })
 					languageTag: sourceLanguageTag,
 					messageId: message.messageId,
 				})
+
+				const translation = resolveEscapedCharacters(translationString)
 
 				const truncatedTranslation =
 					translation &&