+++ inlang/source-code/paraglide/paraglide-js-adapter-next/example-pages/next.config.js
@@ -1,0 +1,10 @@
+import { availableLanguageTags, sourceLanguageTag } from "./src/paraglide/runtime.js"
+
+/** @type {import('next').NextConfig} */
+export default {
+	reactStrictMode: true,
+	i18n: {
+		locales: [...availableLanguageTags],
+		defaultLocale: sourceLanguageTag,
+	},
+}