+++ inlang/source-code/paraglide/paraglide-js/bin/run.js
@@ -1,5 +1,4 @@
 #!/usr/bin/env node
-
-import { cli } from "../dist/cli/main.js"
+import { cli } from "../dist/index.js"
 
 cli.parse()