+++ inlang/source-code/website/src/renderer/_default.page.server.tsx
@@ -55,7 +55,7 @@ export async function render(pageContext: PageContextRenderer): Promise<unknown>
 			${dangerouslySkipEscape(import.meta.env.PROD ? analytics : "")}
 			${dangerouslySkipEscape(favicons)}
 			${dangerouslySkipEscape(generateHydrationScript())}
-			${dangerouslySkipEscape(renderTags(tags))}
+			${languageTag() === "en" || languageTag() === "de" ? dangerouslySkipEscape(renderTags(tags)) : ""}
       </head>
 	  <!-- setting min-h/w-screen to allow child elements to span to the entire screen  -->
       <body class="website min-h-screen min-w-screen bg-background text-on-background" id="root">